Convert "unsigned long" to guint32 in preparation for 64bit machinery.
* app/composite/gimp-composite-mmx.c
  app/composite/gimp-composite-sse.c
  app/composite/gimp-composite-sse2.c:
  Convert "unsigned long" to guint32 in preparation for 64bit
  machinery.

+++ app/composite/gimp-composite-sse2.c
@@ -40,7 +40,7 @@
 #define pminub(src,dst,tmp)  "pminub " "%%" #src ", %%" #dst
 #define pmaxub(src,dst,tmp)  "pmaxub " "%%" #src ", %%" #dst
 
-const static unsigned long rgba8_alpha_mask_128[4] = { 0xFF000000, 0xFF000000, 0xFF000000, 0xFF000000 };
+const static guint32 rgba8_alpha_mask_128[4] = { 0xFF000000, 0xFF000000, 0xFF000000, 0xFF000000 };
 
 void
 debug_display_sse(void)
